delete mok
