+++ app/src/main/java/ru/practicum/android/diploma/data/dto/VacancyDto.kt
@@ -29,61 +29,3 @@ data class VacancyDto(
     val pages: Int,
     val per_page: Int,
 ) : Parcelable
-
-@Parcelize
-data class Area(
-    val id: String,
-    val name: String,
-    val url: String,
-) : Parcelable
-
-@Parcelize
-data class Department(
-    val id: String,
-    val name: String,
-) : Parcelable
-
-@Parcelize
-data class Employer(
-    val accredited_it_employer: Boolean,
-    val alternate_url: String?,
-    val id: String?,
-    val logo_urls: LogoUrls?,
-    val name: String,
-    val trusted: Boolean,
-    val url: String?,
-    val vacancies_url: String?,
-) : Parcelable
-
-@Parcelize
-data class ProfessionalRoles(
-    val id: String,
-    val name: String,
-) : Parcelable
-
-@Parcelize
-data class Salary(
-    val currency: String?,
-    val from: Int?,
-    val gross: Boolean?,
-    val to: Int?,
-) : Parcelable
-
-@Parcelize
-data class Type(
-    val id: String,
-    val name: String,
-) : Parcelable
-
-@Parcelize
-data class Snippet(
-    val requirement: String?,
-    val responsibility: String?,
-) : Parcelable
-
-@Parcelize
-data class LogoUrls(
-    val v90: String,
-    val v240: String,
-    val original: String,
-) : Parcelable
